+++ exonum-light-client/src/main/java/com/exonum/client/ExonumClient.java
@@ -201,6 +201,7 @@ public interface ExonumClient {
 
     private URL exonumHost;
     private OkHttpClient httpClient = DEFAULT_CLIENT;
+    private String prefix = "";
 
     /**
      * Sets Exonum host url.
@@ -236,12 +237,22 @@ public interface ExonumClient {
     }
 
     /**
+     * Sets an optional URL prefix to be applied to all requests made by the client.
+     * Can be helpful in case of using middleware routing proxy on the blockchain node side.
+     * There is no prefix by default.
+     */
+    public Builder setPrefix(String prefix) {
+      this.prefix = checkNotNull(prefix);
+      return this;
+    }
+
+    /**
      * Creates Exonum client instance.
      * @throws IllegalStateException if required fields weren't set
      */
     public ExonumClient build() {
       checkRequiredFieldsSet();
-      return new ExonumHttpClient(httpClient, exonumHost);
+      return new ExonumHttpClient(httpClient, exonumHost, prefix);
     }
 
     private void checkRequiredFieldsSet() {