+++ exonum-light-client/CHANGELOG.md
@@ -16,8 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
-* `ExonumClient#findNonEmptyBlocks` to find a certain number of the most recent non-empty
+- `ExonumClient#findNonEmptyBlocks` to find a certain number of the most recent non-empty
   blocks (from the last block and up to the genesis block). (#953)
+- Prefix URL can be set for routing all Light Client requests. (#997) 
 
 ### Changed
 - `ExonumClient#getBlocks` accepts a closed range of block heights _[from; to]_
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `ExonumClient#getLastBlocks` returns only the blocks from the closed range 
   _[max(0, blockchainHeight - count + 1), blockchainHeight]_. The size of
   the range _(min(count, blockchainHeight + 1))_ is no longer limited. (#953)
+- Now port is optional in the Exonum host URL. (#997) 
 
 ## 0.2.0 - 2019-05-27